+++ lib/nanook/wallet.rb
@@ -149,6 +149,18 @@ class Nanook
       end
     end
 
+    # Changes a wallet's seed.
+    #
+    # The seed is the wallet's unique identifier.
+    #
+    # @param seed [String] the seed/id to change to.
+    # @return [Boolean] indicating whether the change was successful.
+    def change_id(seed)
+      wallet_required!
+      rpc(:wallet_change_seed, seed: seed).has_key?(:success)
+    end
+    alias_method :change_seed, :change_id
+
     # Creates a new wallet.
     #
     #   Nanook.new.wallet.create
@@ -214,6 +226,7 @@ class Nanook
     def id
       @wallet
     end
+    alias_method :seed, :id
 
     def inspect # :nodoc:
       "#{self.class.name}(id: \"#{id}\", object_id: \"#{"0x00%x" % (object_id << 1)}\")"
@@ -258,6 +271,76 @@ class Nanook
       account(from).pay(to: to, amount: amount, unit: unit, id: id)
     end
 
+    # Returns a Hash of accounts in this wallet with their pending block hash ids.
+    # A pending block is a payment that has not yet been receieved by an account
+    # in the wallet.
+    #
+    # See also the #receive method of this class for how to receive a pending payment.
+    #
+    # ==== Arguments
+    #
+    # [+limit:+]    Number of accounts with pending blocks to return (default is 1000)
+    # [+threshold:+] Only return accounts with pendings blocks greater than this amount
+    #               (default is +nil+). When +nil+ this argument will not apply.
+    #               The unit of +threshold+ is determined by the +unit:+ argument and defaults
+    #               to NANO.
+    # [+unit:+]   Symbol (default is +:nano+) Sets the unit of the +threshold:+ argument, and
+    #             also the unit of the response.
+    #             Must be either +:nano+ or +:raw+. (Note: this method
+    #             interprets +:nano+ as NANO, which is technically Mnano
+    #             See {What are Nano's Units}[https://nano.org/en/faq#what-are-nano-units-])
+    #
+    # ==== Example 1
+    #
+    #   wallet.pending
+    #
+    # ==== Example 1 response
+    #   {
+    #     "xrb_1111111111111111111111111111111111111111111111111117353trpda"=>["142A538F36833D1CC78B94E11C766F75818F8B940771335C6C1B8AB880C5BB1D","718CC2121C3E641059BC1C2CFC45666C99E8AE922F7A807B7D07B62C995D79E2"],
+    #     "xrb_3t6k35gi95xu6tergt6p69ck76ogmitsa8mnijtpxm9fkcm736xtoncuohr3"=>["4C1FEEF0BEA7F50BE35489A1233FE002B212DEA554B55B1B470D78BD8F210C74"]
+    #   }
+    # ==== Example 2
+    #
+    #   wallet.pending(threshold: 1)
+    #
+    # ==== Example 2 response
+    #   TODO
+    #
+    # ==== Example 3
+    #
+    #   wallet.pending(threshold: 999999999999999999999999, unit: :raw)
+    #
+    # ==== Example 3 response
+    #   TODO
+    def pending(limit: 1000, threshold:nil, unit: Nanook::WalletAccount::DEFAULT_UNIT)
+      wallet_required!
+
+      unless Nanook::WalletAccount::UNITS.include?(unit)
+        raise ArgumentError.new("Unsupported unit: #{unit}")
+      end
+
+      if !threshold.nil? && unit.eql?(:nano)
+        threshold = Nanook::Util.NANO_to_raw(threshold)
+      end
+
+      params = {
+        count: limit
+      }
+
+      if !threshold.nil?
+        params.merge!(threshold: threshold)
+      end
+
+      response = rpc(:wallet_pending, params)[:blocks]
+
+      if !threshold.nil?
+        # response = response.map do |account, blocks|
+        # end
+      end
+
+      Nanook::Util.coerce_empty_string_to_type(response, Hash)
+    end
+
     # Receives a pending payment into an account in the wallet.
     #
     # When called with no +block+ argument, the latest pending payment
@@ -292,6 +375,33 @@ class Nanook
       wallet_required!
       validate_wallet_contains_account!(into)
       account(into).receive(block)
+    end
+
+    # Restore a previously created wallet on this node.
+    #
+    # ==== Example:
+    #
+    #   Nanook.new.wallet.restore(wallet_id) # => "718CC2121C3E641059BC1C2CFC45666C99E8AE922F7A807B7D07B62C995D79E2"
+    #
+    # @param wallet [String] the wallet id to restore.
+    # @param accounts [Integer] optionally restore the given number of accounts for the wallet.
+    #
+    # @return [String] the new wallet id
+    # @raise Nanook::Error if unsuccessful
+    def restore(wallet, accounts:0)
+      @wallet = create
+
+      unless change_id(wallet)
+        raise Nanook::Error.new("Unable to set seed/id for wallet")
+      end
+
+      @wallet = wallet
+
+      if accounts > 0
+        account.create(accounts)
+      end
+
+      @wallet
     end
 
     # Returns a boolean to indicate if the wallet is locked.