+++ spec/nanook/wallet_spec.rb
@@ -4,14 +4,6 @@ RSpec.describe Nanook::Wallet do
   let(:account_id) { "xrb_3e3j5tkog48pnny9dmfzj1r16pg8t1e76dz5tmac6iq689wyjfpi00000000" }
   let(:wallet_id) { "000D1BAEC8EC208142C99059B393051BAC8380F9B5A2E6B2489A277D81789F3F" }
   let(:block_id) { "000D1BAEC8EC208142C99059B393051BAC8380F9B5A2E6B2489A277D81789F3F" }
-  let(:headers) {
-    {
-      'Accept'=>'*/*',
-      'Accept-Encoding'=>'gzip;q=1.0,deflate;q=0.6,identity;q=0.3',
-      'Content-Type'=>'application/json',
-      'User-Agent'=>'Ruby nanook gem'
-    }
-  }
 
   def stub_valid_account_check
     stub_request(:post, "http://localhost:7076/").
@@ -178,16 +170,19 @@ RSpec.describe Nanook::Wallet do
   it "wallet send payment" do
     stub_valid_account_check
 
-    stub_request(:post, "http://localhost:7076/").
-    with(
-      body: "{\"action\":\"validate_account_number\",\"account\":\"xrb_3e3j5tkog48pnny9dmfzj1r16pg8t1e76dz5tmac6iq689wyjfpi00000000\"}",
-      headers: {
-      'Accept'=>'*/*',
-      'Accept-Encoding'=>'gzip;q=1.0,deflate;q=0.6,identity;q=0.3',
-      'Content-Type'=>'application/json',
-      'User-Agent'=>'Ruby nanook gem'
-      }).
-    to_return(status: 200, body: "{\"valid\":\"1\"}", headers: {})
+    stub_request(:post, uri).with(
+      body: "{\"action\":\"account_info\",\"account\":\"#{account_id}\"}",
+      headers: headers
+    ).to_return(
+      status: 200,
+      body: "{\"frontier\":\"FF84533A571D953A596EA401FD41743AC85D04F406E76FDE4408EAED50B473C5\",
+      \"open_block\":\"991CF190094C00F0B68E2E5F75F6BEE95A2E0BD93CEAA4A6734DB9F19B728948\",
+      \"representative_block\":\"991CF190094C00F0B68E2E5F75F6BEE95A2E0BD93CEAA4A6734DB9F19B728948\",
+      \"balance\":\"23\",
+      \"modified_timestamp\":\"1501793775\",
+      \"block_count\":\"33\"}",
+      headers: {}
+    )
 
     stub_request(:post, uri).with(
       body: "{\"action\":\"send\",\"wallet\":\"#{wallet_id}\",\"source\":\"#{account_id}\",\"destination\":\"#{account_id}\",\"amount\":\"2000000000000000000000000000000\",\"id\":\"7081e2b8fec9146e\"}",
@@ -205,16 +200,19 @@ RSpec.describe Nanook::Wallet do
   it "wallet send payment in raw" do
     stub_valid_account_check
 
-    stub_request(:post, "http://localhost:7076/").
-    with(
-      body: "{\"action\":\"validate_account_number\",\"account\":\"xrb_3e3j5tkog48pnny9dmfzj1r16pg8t1e76dz5tmac6iq689wyjfpi00000000\"}",
-      headers: {
-      'Accept'=>'*/*',
-      'Accept-Encoding'=>'gzip;q=1.0,deflate;q=0.6,identity;q=0.3',
-      'Content-Type'=>'application/json',
-      'User-Agent'=>'Ruby nanook gem'
-      }).
-    to_return(status: 200, body: "{\"valid\":\"1\"}", headers: {})
+    stub_request(:post, uri).with(
+      body: "{\"action\":\"account_info\",\"account\":\"#{account_id}\"}",
+      headers: headers
+    ).to_return(
+      status: 200,
+      body: "{\"frontier\":\"FF84533A571D953A596EA401FD41743AC85D04F406E76FDE4408EAED50B473C5\",
+      \"open_block\":\"991CF190094C00F0B68E2E5F75F6BEE95A2E0BD93CEAA4A6734DB9F19B728948\",
+      \"representative_block\":\"991CF190094C00F0B68E2E5F75F6BEE95A2E0BD93CEAA4A6734DB9F19B728948\",
+      \"balance\":\"23\",
+      \"modified_timestamp\":\"1501793775\",
+      \"block_count\":\"33\"}",
+      headers: {}
+    )
 
     stub_request(:post, uri).with(
       body: "{\"action\":\"send\",\"wallet\":\"#{wallet_id}\",\"source\":\"#{account_id}\",\"destination\":\"#{account_id}\",\"amount\":\"2\",\"id\":\"7081e2b8fec9146e\"}",
@@ -357,4 +355,126 @@ RSpec.describe Nanook::Wallet do
     expect(response[:pending]).to eq(2000000000000000000000000000)
   end
 
+  it "wallet restore with no accounts" do
+    initial_wallet_id = "000F2BAEC8EC208142C99059B393051BAC8380F9B5A2E6B2489A277D81789F3F"
+
+    stub_request(:post, uri).with(
+      body: "{\"action\":\"wallet_create\"}",
+      headers: headers
+    ).to_return(
+      status: 200,
+      body: "{\"wallet\":\"#{initial_wallet_id}\"}",
+      headers: {}
+    )
+
+    stub_request(:post, uri).with(
+      body: "{\"action\":\"wallet_change_seed\",\"wallet\":\"#{initial_wallet_id}\",\"seed\":\"#{wallet_id}\"}",
+      headers: headers
+    ).to_return(
+      status: 200,
+      body: "{\"success\":\"\"}",
+      headers: {}
+    )
+
+    response = Nanook.new.wallet.restore(wallet_id)
+    expect(response).to eq wallet_id
+  end
+
+  it "wallet restore with 2 accounts" do
+    initial_wallet_id = "000F2BAEC8EC208142C99059B393051BAC8380F9B5A2E6B2489A277D81789F3F"
+
+    stub_request(:post, uri).with(
+      body: "{\"action\":\"wallet_create\"}",
+      headers: headers
+    ).to_return(
+      status: 200,
+      body: "{\"wallet\":\"#{initial_wallet_id}\"}",
+      headers: {}
+    )
+
+    stub_request(:post, uri).with(
+      body: "{\"action\":\"wallet_change_seed\",\"wallet\":\"#{initial_wallet_id}\",\"seed\":\"#{wallet_id}\"}",
+      headers: headers
+    ).to_return(
+      status: 200,
+      body: "{\"success\":\"\"}",
+      headers: {}
+    )
+
+    stub_request(:post, uri).with(
+      body: "{\"action\":\"account_create\",\"wallet\":\"#{wallet_id}\"}",
+      headers: headers
+    ).to_return(
+      status: 200,
+      body: "{\"account\":\"xrb_3e3j5tkog48pnny9dmfzj1r16pg8t1e76dz5tmac6iq689wyjfpi00000000\"}",
+      headers: {}
+    )
+
+    response = Nanook.new.wallet.restore(wallet_id)
+    expect(response).to eq wallet_id
+  end
+
+  it "wallet pending with default limit" do
+    stub_request(:post, uri).with(
+      body: "{\"action\":\"wallet_pending\",\"wallet\":\"#{wallet_id}\",\"count\":\"1000\"}",
+      headers: headers
+    ).to_return(
+      status: 200,
+      body: "{\"blocks\" : {
+        \"xrb_1111111111111111111111111111111111111111111111111117353trpda\": [\"142A538F36833D1CC78B94E11C766F75818F8B940771335C6C1B8AB880C5BB1D\",\"142A538F36833D1CC78B94E11C766F75818F8B940771335C6C1B8AB880C5BB1D\"],
+        \"xrb_3t6k35gi95xu6tergt6p69ck76ogmitsa8mnijtpxm9fkcm736xtoncuohr3\": [\"4C1FEEF0BEA7F50BE35489A1233FE002B212DEA554B55B1B470D78BD8F210C74\"]
+      }}",
+      headers: {}
+    )
+
+    response = Nanook.new.wallet(wallet_id).pending
+    expect(response.keys).to have(2).items
+    expect(response["xrb_1111111111111111111111111111111111111111111111111117353trpda"]).to have(2).items
+    expect(response["xrb_3t6k35gi95xu6tergt6p69ck76ogmitsa8mnijtpxm9fkcm736xtoncuohr3"]).to have(1).item
+  end
+
+  it "wallet pending with limit" do
+    stub_request(:post, uri).with(
+      body: "{\"action\":\"wallet_pending\",\"wallet\":\"#{wallet_id}\",\"count\":\"1\"}",
+      headers: headers
+    ).to_return(
+      status: 200,
+      body: "{\"blocks\" : {
+        \"xrb_1111111111111111111111111111111111111111111111111117353trpda\": [\"142A538F36833D1CC78B94E11C766F75818F8B940771335C6C1B8AB880C5BB1D\",\"142A538F36833D1CC78B94E11C766F75818F8B940771335C6C1B8AB880C5BB1D\"]
+      }}",
+      headers: {}
+    )
+
+    response = Nanook.new.wallet(wallet_id).pending(limit: 1)
+    expect(response.keys).to have(1).items
+    expect(response["xrb_1111111111111111111111111111111111111111111111111117353trpda"]).to have(2).items
+  end
+
+  it "wallet pending with threshold" do
+    skip
+    stub_request(:post, uri).with(
+      body: "{\"action\":\"wallet_pending\",\"wallet\":\"#{wallet_id}\",\"count\":\"1\",\"threshold\":\"1000000000000000000000000000000\"}",
+      headers: headers
+    ).to_return(
+      status: 200,
+      body: "{\"blocks\": {
+        \"xrb_1111111111111111111111111111111111111111111111111117353trpda\": {
+            \"142A538F36833D1CC78B94E11C766F75818F8B940771335C6C1B8AB880C5BB1D\": \"6000000000000000000000000000000\",
+            \"242A538F36833D1CC78B94E11C766F75818F8B940771335C6C1B8AB880C5BB1D\": \"34524565367345234523452344356356745674\"
+        },
+        \"xrb_3t6k35gi95xu6tergt6p69ck76ogmitsa8mnijtpxm9fkcm736xtoncuohr3\": {
+            \"4C1FEEF0BEA7F50BE35489A1233FE002B212DEA554B55B1B470D78BD8F210C74\": \"106370018000000000000000000000000\"
+        }
+      }",
+      headers: {}
+    )
+
+    response = Nanook.new.wallet(wallet_id).pending(limit: 1, threshold: 1)
+    expect(response.keys).to have(1).items
+    expect(response["xrb_1111111111111111111111111111111111111111111111111117353trpda"].keys).to have(1).item
+    expect(response["xrb_1111111111111111111111111111111111111111111111111117353trpda"]["142A538F36833D1CC78B94E11C766F75818F8B940771335C6C1B8AB880C5BB1D"]).to eql 6
+  end
+
+
+  it "wallet pending with threshold and raw unit"
 end